Made row_type be uint64_t instead of uint32_t

+++ src/board.hh
@@ -40,7 +40,7 @@ class Board {
   
 public:
   using dim_type = unsigned short;
-  using row_type = uint32_t;
+  using row_type = uint64_t;
 
   static constexpr dim_type BOARD_DIM = 8;
